macOS Apple Silicon (arm64) build support
- CMakeLists.txt: force C mode for .C sources under AppleClang (which
  otherwise treats uppercase .C as C++); guard OpenAL include dirs
  against an empty source dir (system framework on macOS).
- include/psx/libgpu.h: include __aarch64__/_M_ARM64 in the 64-bit
  extended-prim-pointer arch macro.
- src/gte/half_float.cpp: cast FP16 init to unsigned short (narrowing
  init fix).
- src/psx/libcd.c: exclude <malloc.h> on Apple (not available).

+++ src/gte/half_float.cpp
@@ -55,7 +55,7 @@ short to_half_float(const float x)
 
 float from_half_float(const short x)
 {
-	FP16 h = { x };
+	FP16 h = { (unsigned short)x };
 
 	static const FP32 magic = { 113 << 23 };
 	static const uint shifted_exp = 0x7c00 << 13; // exponent mask after shift
